Fixed HomeScreenViewModel
