+++ trackr-app/app/src/main/java/com/trackr/trackr_app/MainActivity.kt
@@ -13,32 +13,17 @@ import com.trackr.trackr_app.ui.theme.TrackrappTheme
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
 
-class MainActivityViewModel: ViewModel() {
-    //NOTE: This design is inspired by documentation from Google:
-    //https://developer.android.com/codelabs/jetpack-compose-state#3
-    private var _events = MutableLiveData(listOf<String>())
-    val events: LiveData<List<String>> = _events
-    fun addEvent(event: String) {
-        _events.value = _events.value!! + listOf(event)
-    }
-}
+
 
 class MainActivity : ComponentActivity() {
-    private val mainActivityViewModel by viewModels<MainActivityViewModel>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
         setContent {
             TrackrappTheme {
-                MainActivityScreen(mainActivityViewModel)
+                MainScreen()
             }
         }
     }
-}
-
-@Composable
-private fun MainActivityScreen(viewModel: MainActivityViewModel) {
-    val events: List<String> by viewModel.events.observeAsState(listOf())
-    MainScreen(eventList = events, onAddItem = {viewModel.addEvent(it)})
 }
