+++ trackr-app/app/src/main/java/com/trackr/trackr_app/ui/home/HomeScreen.kt
@@ -1,13 +1,18 @@
 package com.trackr.trackr_app.ui.home
 
 import android.content.res.Configuration.UI_MODE_NIGHT_YES
+import android.util.Log
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.*
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Add
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Brush
@@ -18,29 +23,58 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
 import com.trackr.trackr_app.R
+import com.trackr.trackr_app.ui.main.MainScreen
 import com.trackr.trackr_app.ui.theme.Rubik
 import com.trackr.trackr_app.ui.theme.TrackrappTheme
 
 
 val temp_data = listOf(1,2,3,4,5)
+//
+//@Preview(
+//    uiMode = UI_MODE_NIGHT_YES  // enables dark mode for preview
+//)
+//@Composable
+//private fun PreviewWithTheme() {
+//    TrackrappTheme {
+//        HomeScreen(listOf<String>("hello", "hi"))
+//    }
+//}
 
-@Preview(
-    uiMode = UI_MODE_NIGHT_YES  // enables dark mode for preview
-)
-@Composable
-private fun PreviewWithTheme() {
-    TrackrappTheme {
-        HomeScreen(listOf<String>("hello", "hi"))
+class HomeScreenViewModel: ViewModel() {
+    //NOTE: This design is inspired by documentation from Google:
+    //https://developer.android.com/codelabs/jetpack-compose-state#3
+    private var _events = MutableLiveData(listOf<String>())
+    val events: LiveData<List<String>> = _events
+    fun addEvent(event: String) {
+        _events.value = _events.value!! + listOf(event)
     }
 }
 
 @Composable
+fun HomeScreenActivity(viewModel: HomeScreenViewModel) {
+    val events: List<String> by viewModel.events.observeAsState(listOf())
+    HomeScreen(eventList = events, onAddItem = {viewModel.addEvent(it)})
+}
+
+@Composable
 fun HomeScreen(
-//    homeViewModel: HomeViewModel,
-    eventList: List<String>
+    eventList: List<String>, onAddItem: (String) -> Unit
 ) {
     Scaffold(
+        floatingActionButton = {
+            FloatingActionButton(
+                onClick = {
+                    onAddItem("Bob's Birthday")
+                    Log.d("HELLO", eventList.toString())
+                },
+                backgroundColor = Color.Blue) {
+                Icon(Icons.Filled.Add, "Add event")
+            }
+        },
         backgroundColor = MaterialTheme.colors.background,
         bottomBar = { BottomAppBar() }
     ) {
