+++ trackr-app/app/src/main/java/com/trackr/trackr_app/ui/main/MainScreen.kt
@@ -1,6 +1,7 @@
 package com.trackr.trackr_app.ui.main
 
 import android.util.Log
+import androidx.activity.viewModels
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material.*
 import androidx.compose.material.icons.Icons
@@ -13,30 +14,25 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.trackr.trackr_app.R
 import com.trackr.trackr_app.entities.Event
+import com.trackr.trackr_app.ui.add.AddScreenActivity
 import com.trackr.trackr_app.ui.home.HomeScreen
+import com.trackr.trackr_app.ui.home.HomeScreenActivity
+import com.trackr.trackr_app.ui.home.HomeScreenViewModel
 import com.trackr.trackr_app.ui.navigation.NavScreen
 import com.trackr.trackr_app.ui.theme.Rubik
 
 @Composable
-fun MainScreen(eventList: List<String>, onAddItem: (String) -> Unit) {
+fun MainScreen() {
+    val homeScreenViewModel: HomeScreenViewModel = viewModel()
     val navController = rememberNavController()
     //val eventList = ArrayList<String>()
     Scaffold(
-        floatingActionButton = {
-            FloatingActionButton(
-                onClick = {
-                    onAddItem("Bob's Birthday")
-                    Log.d("HELLO", eventList.toString())
-                          },
-                backgroundColor = Color.Blue) {
-                Icon(Icons.Filled.Add, "Add event")
-            }
-        },
         topBar = {
             TopAppBar(
                 backgroundColor = MaterialTheme.colors.primary
@@ -53,7 +49,10 @@ fun MainScreen(eventList: List<String>, onAddItem: (String) -> Unit) {
         }
     ) {
         NavHost(navController = navController, startDestination = NavScreen.Home.route) {
-            composable(NavScreen.Home.route) { HomeScreen(eventList = eventList) }
+            composable(NavScreen.Home.route) { HomeScreenActivity(homeScreenViewModel) }
         }
+//        NavHost(navController = navController, startDestination = NavScreen.Home.route) {
+//            composable(NavScreen.Home.route) { AddScreenActivity(homeScreenViewModel) }
+//        }
     }
 }
